Oct 23,2022

+++ com.SportyShoes/src/main/java/com/SportyShoes/Repository/UserRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.SportyShoes.Entity.Users;
 
 @Repository
-public interface UserRepo extends CrudRepository<Users, Integer> {
+public interface UserRepo extends CrudRepository<Users, Long> {
 	
 	@Query(value="select * from Users u where u.firstname like %:keyword% or u.lastname like %:keyword% or u.email like %:keyword%",nativeQuery = true)
 	public List<Users> findbykeyword(@Param("keyword") String keyword);
